requirements.txt
byteplus-python-sdk-v2
httpx
pydantic
sniffio
Pillow
numpy
requests